// src/features/user/types/PeriodStatByUser.ts
export type PeriodStatByUser = {
  boost_ap_2_bonus: number;
  boost_m1_1_bonus: number;
  infinity_bonus: number;
  lo: number;
  cashback_total_800_bonus: number;
  boost_m2_1_bonus: number;
  invite_bonus: number;
  cashback_bonus: number;
  premium: number;
  ob: number;
  fix: number;
  client: number;
  state: string;
  id: string;
  boost_gm_2_bonus: number;
  tour_bonus: number;
  prime_bonus: number;
  cashback_total_5000_bonus: number;
  cashback_total_2000_bonus: number;
  boost_ap_1_bonus: number;
  newbie: number;
  prime_bonus_info: string;
  go: number;
  team_bonus: number;
  cashback_bonus_info: string;
  active: number;
  boost_m2_2_bonus: number;
  boost_m1_2_bonus: number;
  lifestyle_bonus: number;
  dynamics_bonus: number;
  qual_name: string;
  nlo: number;
  invite_bonus_info: string;
  team_bonus_info: string;
  infinity_bonus_info: string;
  premium_gift: number;
  dynamics_bonus_info: string;
  boost_gm_1_bonus: number;
  qual: number;
  start_bonus: number;
  user: {
    birthday: string;
    gender: string;
    timezone: string;
    settlement_fias_id: string;
    pat_name: string;
    last_name: string;
    telegram: string;
    instagram: string;
    login: string;
    delete: number;
    settlement: string;
    co_statement_link: string;
    phone: string;
    date_create: string;
    theme: string;
    id: string;
    first_name: string;
    payout_state: string;
    email: string;
    zulip_api_key: string;
    restore_code: string;
    state: string;
    state_name: string;
  };
  children_user: number;
  level: number;
  patch_list: unknown[];
};